+++ CHANGELOG.md
@@ -10,17 +10,18 @@ Additions:
 - Added attributes property to various tables
 - Added is_valid property to Verification
 
+Removals:
+- Remove html from Article and Product (use attributes instead)
+
 Improvements:
 - Invoices now have references back to the order
 - Updated various foreign key on delete actions
 - Relationships with an order attribute will be ordered by default
 
 Todo:
-- Remove html from Article and Product, instead use attributes
 - Rename desc to description in all tables
 - Rename img_url to image_url in all tables
 - Rename read_html to consent_required in all tables
-- Refactor Invoice, Refund and Shipment under Order
 
 ### 2023-07-04